[ALPS02020409]WMT/STP:fix KE issue when stack protector is enabled
[Detail]32bit data should not be used in wmt_core_ctrl function

[Solution]wmt_core_ctrl parameter type change to unsigned long

[Feature] MTK BT Solution(*)

MTK-Commit-Id: 9754378235c05e05cbcd427ce827440613313fe4

CR-Id: ALPS02020409

+++ drivers/misc/mediatek/connectivity/conn_soc/common/core/wmt_func.c
@@ -270,8 +270,8 @@ INT32 wmt_func_bt_on(P_WMT_IC_OPS pOps, P_WMT_GEN_CONF pConf)
 {
 	/* return wmt_func_bt_ctrl(FUNC_ON); */
 	INT32 iRet = -1;
-	UINT32 ctrlPa1;
-	UINT32 ctrlPa2;
+	unsigned long ctrlPa1;
+	unsigned long ctrlPa2;
 
 	ctrlPa1 = BT_PALDO;
 	ctrlPa2 = PALDO_ON;
@@ -309,8 +309,8 @@ INT32 wmt_func_bt_off(P_WMT_IC_OPS pOps, P_WMT_GEN_CONF pConf)
 	/* return wmt_func_bt_ctrl(FUNC_OFF); */
 	INT32 iRet1 = -1;
 	INT32 iRet2 = -1;
-	UINT32 ctrlPa1;
-	UINT32 ctrlPa2;
+	unsigned long ctrlPa1;
+	unsigned long ctrlPa2;
 
 	iRet1 = wmt_core_func_ctrl_cmd(WMTDRV_TYPE_BT, MTK_WCN_BOOL_FALSE);
 	if (iRet1)
@@ -483,8 +483,8 @@ INT32 wmt_func_gps_pre_off(P_WMT_IC_OPS pOps, P_WMT_GEN_CONF pConf)
 INT32 wmt_func_gps_on(P_WMT_IC_OPS pOps, P_WMT_GEN_CONF pConf)
 {
 	INT32 iRet = 0;
-	UINT32 ctrlPa1;
-	UINT32 ctrlPa2;
+	unsigned long ctrlPa1;
+	unsigned long ctrlPa2;
 	UINT8 co_clock_type = (pConf->co_clock_flag & 0x0f);
 
 	if ((co_clock_type) && (0 == pConf->wmt_gps_lna_enable)) {	/* use SOC external LNA */
@@ -520,8 +520,8 @@ INT32 wmt_func_gps_on(P_WMT_IC_OPS pOps, P_WMT_GEN_CONF pConf)
 INT32 wmt_func_gps_off(P_WMT_IC_OPS pOps, P_WMT_GEN_CONF pConf)
 {
 	INT32 iRet = 0;
-	UINT32 ctrlPa1 = 0;
-	UINT32 ctrlPa2 = 0;
+	unsigned long ctrlPa1 = 0;
+	unsigned long ctrlPa2 = 0;
 	UINT8 co_clock_type = (pConf->co_clock_flag & 0x0f);
 
 	iRet = wmt_func_gps_pre_off(pOps, pConf);
@@ -567,8 +567,8 @@ INT32 _osal_inline_ wmt_func_fm_ctrl(ENUM_FUNC_STATE funcState)
 INT32 wmt_func_fm_on(P_WMT_IC_OPS pOps, P_WMT_GEN_CONF pConf)
 {
 	/* return wmt_func_fm_ctrl(FUNC_ON); */
-	UINT32 ctrlPa1 = 0;
-	UINT32 ctrlPa2 = 0;
+	unsigned long ctrlPa1 = 0;
+	unsigned long ctrlPa2 = 0;
 	INT32 iRet = -1;
 	UINT8 co_clock_type = (pConf->co_clock_flag & 0x0f);
 
@@ -594,8 +594,8 @@ INT32 wmt_func_fm_on(P_WMT_IC_OPS pOps, P_WMT_GEN_CONF pConf)
 INT32 wmt_func_fm_off(P_WMT_IC_OPS pOps, P_WMT_GEN_CONF pConf)
 {
 	/* return wmt_func_fm_ctrl(FUNC_OFF); */
-	UINT32 ctrlPa1 = 0;
-	UINT32 ctrlPa2 = 0;
+	unsigned long ctrlPa1 = 0;
+	unsigned long ctrlPa2 = 0;
 	INT32 iRet = -1;
 	UINT8 co_clock_type = (pConf->co_clock_flag & 0x0f);
 
@@ -625,8 +625,8 @@ INT32 wmt_func_fm_off(P_WMT_IC_OPS pOps, P_WMT_GEN_CONF pConf)
 INT32 wmt_func_wifi_ctrl(ENUM_FUNC_STATE funcState)
 {
 	INT32 iRet = 0;
-	UINT32 ctrlPa1 = WMT_SDIO_FUNC_WIFI;
-	UINT32 ctrlPa2 = (FUNC_ON == funcState) ? 1 : 0;	/* turn on Wi-Fi driver */
+	unsigned long ctrlPa1 = WMT_SDIO_FUNC_WIFI;
+	unsigned long ctrlPa2 = (FUNC_ON == funcState) ? 1 : 0;	/* turn on Wi-Fi driver */
 
 	iRet = wmt_core_ctrl(WMT_CTRL_SDIO_FUNC, &ctrlPa1, &ctrlPa2);
 	if (iRet) {
@@ -640,8 +640,8 @@ INT32 wmt_func_wifi_ctrl(ENUM_FUNC_STATE funcState)
 INT32 wmt_func_wifi_on(P_WMT_IC_OPS pOps, P_WMT_GEN_CONF pConf)
 {
 	int iRet = 0;
-	UINT32 ctrlPa1;
-	UINT32 ctrlPa2;
+	unsigned long ctrlPa1;
+	unsigned long ctrlPa2;
 
 	if (NULL != mtk_wcn_wlan_probe) {
 
@@ -678,8 +678,8 @@ INT32 wmt_func_wifi_off(P_WMT_IC_OPS pOps, P_WMT_GEN_CONF pConf)
 {
 	int iRet = 0;
 
-	UINT32 ctrlPa1 = 0;
-	UINT32 ctrlPa2 = 0;
+	unsigned long ctrlPa1 = 0;
+	unsigned long ctrlPa2 = 0;
 
 	if (NULL != mtk_wcn_wlan_remove) {
